Factor tt::is_callable into its own file
