+++ src/NumericalAlgorithms/DiscontinuousGalerkin/Protocols.hpp
@@ -8,6 +8,7 @@
 #include "Utilities/TypeTraits.hpp"
 #include "Utilities/TypeTraits/CreateHasTypeAlias.hpp"
 #include "Utilities/TypeTraits/CreateIsCallable.hpp"
+#include "Utilities/TypeTraits/IsCallable.hpp"
 
 namespace dg {
 /// \ref protocols related to Discontinuous Galerkin functionality
